made changes to hdwallet and incorporated into bip32
